Incorporate recency into catalog ranker (#266)
This changeset introduces a new option for incorporating the recency of the last update timestamp
into the scoring of datasets in Cetera. Specifically, there is a new query parameter `age_decay`
that can be passed to the catalog endpoint in order to demote older datasets in the search
results.

Format: age_decay=={DECAY_TYPE},{SCALE},{DECAY},{OFFSET}[,{ORIGIN}]

See the Elasticsearch documentation for more information on each of these parameters:

https://www.elastic.co/guide/en/elasticsearch/reference/current/query-dsl-function-score-query.html#function-decay

Examples:

```sh
curl -XGET localhost:5704/catalog/v1 q==schools age_decay=gauss,365d,0.5,14d
curl -XGET localhost:5704/catalog/v1 q==schools age_decay=gauss,182d,0.5,14d,2017-04-05
```

Additionally, if we want to enable age decay by default, we can include the desired
settings in our service configuration file.

Example:

```
  elasticsearch {
    es-server = localhost
    es-cluster-name = es_dev
    min-should-match = 3<60%
    title-boost = 5.0
    function-score-scripts = ["views", "score"]
    index-alias-name = catalog
    age-decay = {
      decay-type = gauss
      scale = 365d
      decay = 0.5
      offset = 14d
    }
  }
```

Offset is not honored if specified in the configuration since in practice, we will never want a
hard-coded origin date for the lifetime of the service. In the absence of a default value, origin
is always set to the current timestamp.

Rather than simply exposing the new age decay functionality as application-level configuration
settings, I opted for query parameters for a few reasons. First and foremost, being able to
experiment with different values for the decay function is convenient and powerful. Having to
restart the service in order to change these parameters is an anti-pattern. Imagine if we wanted to
A/B test a few different settings; this is achievable when we expose them as query parameters to
the HTTP service, whereas it would be extremely difficult otherwise. Additionally, introducing
these parameters as application-level configuration settings would mean that both the ODN and
customer catalogs would be limited to the same default parameters (barring some added business
logic in the Cetera).

This code has been tested locally on prod-like data. Unit tests pass.

Reviewed-by: ayn.leslie