EN-15151 Error out with helpful message when offset + limit exceeds value of max_result_window (#274)
Elasticsearch has an index configuration parameter named "max_result_window". When the sum of limit
and offset exceeds that value, Cetera 500s. This change updates Cetera to return a helpful error
message when that situation arises, which is rare.

We will want to support the rare cases when deep scrolling on large result sets (more than 10000
results). But that is outside of the scope of this changeset.

Reviewed-by: ayn.leslie, josh.oberwetter