Ensure that column name searches are case insensitive (#277)
With this change, we compose a sequence of match queries as bool should constraints when constructing
a column name filter. Previously, we were using a `terms` query which -- an important point to
remember -- does not use analysis (and was thus not matching on case variants).

Unit tests pass. Tested locally against prod-like index.