EN-16364 Ignore non alphanumeric characters and case when sorting datasets by name (#279)
* EN-16364 Ignore non alphanumeric characters and case when sorting datasets by name

With this changeset, we add a new subfield to the dataset title field (`indexed_metadata.name`)
that uses an analyzer that

1. removes non-alphanumeric characters
2. normalizes whitespace, and
3. lowercases

This enables us to sort on title without having dataset names that start with special characters
bubble to the top. It also allows us to do case insensitive sorting.

There is a corresponding change in Rammstein.

Unit tests pass. Tested locally on prod-like index.

* Check first and last results in sort tests

Reviewed-by: ayn.leslie