EN-16330 Add owner name to results and make owner name sortable (#283)
To return the owner name along with catalog search results, we denormalize the
index even further. In addition to owner user ID, display name and email address are now reproduced
in the document index. (Unlike display names and IDs, we do not actually return email address to API users.) This allows us to a) return them (except for email address) and b) sort on them.

Unit tests pass. Tested locally on prod-like index.

* Fix broken tests and style check

* EN-16502 Ensure datatype and domain category are sortable

* Fix test after rebasing

* Remove "type": "object" from new document.owner field in mapping

* Remove unnecessary describeOwner param and ensure we don't leak owner email

* Address PR comments

Reviewed-by: ayn.leslie