Analyze numbers correctly in lowercase alphanumeric analyzer (#286)
* Analyze numbers correctly in lowercase alphanumeric analyzer

We recently added the lowercase alphanumeric analyzer to have more intuitive sorting on a number
of sortable fields in the new Asset Manager. Specifically, we wanted to ensure that

- special characters are ignored
- case is ignored

But the lowercase alphanumeric analyzer that we added was not handling numbers correctly -- it was
removing them altogether, which would result in unexpected behavior for fields with numbers in
them.

This change makes a small modification to the lowercase alphanumeric analyzer definition in the
ES settings JSON file. It also creates a new AnalyzerSpec test suite to test analysis expections
changes directly.

Unit tests pass. Tested locally on prod-like index. There is a corresponding Rammstein change.

* Add test case for initial and terminal non-alphanumeric characters

Reviewed-by: ayn.leslie