Remove deprecated owner field (#287)
* Remove deprecated owner ID field

This accompanies the following Rammstein PR:

https://github.com/socrata/rammstein/pull/236

* Remove deprecated ownerId field

We were still referring to the old deprecated ownerId field in some places. This has been deprecated
with the new owner object field. This updates all of the test documents and unit tests to use the
owner field. It also replaces a lingering reference in the Document model.

All unit tests pass. Tested locally on prod-like index.

Reviewed-by: ayn.leslie