EN-18636 Add API to support user autocomplete (#303)
* EN-18636 Add API to support user autocomplete

This provides a new endpoint to get user suggestions. It uses the same analyzers as the existing
catalog autocomplete. There are a few intentional changes, however. Firstly, since users don't
suffer from the same problem of duplicates that we have with asset titles, there is no need to
dedupe. As a result, we can return the actual suggest user object along w/ the match spans to
enable highlighting. Additionally, we include "match fields". Since we want to enable autocomplete
on both email and screen name, if we wish to support highlighting, we need to indicate which fields
were the basis for the match.

Adding a UserAutocompleteService -- and the corresponding suite of tests -- results in quite a bit
of redundancy. There may be a good way to remove some of that redundancy, but it would require
restructuring and refactoring a bit of code, and I deemed it out of scope for this chunk of work.

Tested in staging.

* Remove Option(source) in EsUser auxiliary constructor

Per discussion w/ @aynleslie, the auxiliary constructor for the EsUser model should not Option-ify
the input string to handle the potential of a null input. Rather than Optioning the input, we
fallback on the existing try/catch blocks that are already in place where the EsUser is constructed
from a JSON string (in the UserClient, and in relevant tests).

I also renamed that apply method to fromSource, since method overloading in general is frowned
upon in Scala, and since there was no good reason to elide the default apply method in favor of the
string parsing auxiliary constructor.

* Make AutocompleteService more consistent w/ structure of UserAutocompleteService

Additionally, remove unnecessary BaseDocumentClient and BaseDomainClient traits.

Reviewed-by: katie.vedder